xtask: add the detector.rs split's sibling files
Completes the previous commit: the mod/use wiring and the six sibling
files (detectors_structural, detectors_flags, detectors_misc,
calibration, render, commands) didn't make it into the rename commit.
