+++ xtask/src/detector/mod.rs
@@ -33,6 +33,20 @@ use mandible_core::{CommandNode, Entity, Provenance, Source};
 use std::collections::BTreeMap;
 use std::path::Path;
 
+mod calibration;
+mod commands;
+mod detectors_flags;
+mod detectors_misc;
+mod detectors_structural;
+mod render;
+
+pub(crate) use calibration::*;
+pub(crate) use commands::*;
+pub(crate) use detectors_flags::*;
+pub(crate) use detectors_misc::*;
+pub(crate) use detectors_structural::*;
+pub(crate) use render::*;
+
 /// Everything a detector is allowed to look at for one tool: the raw help
 /// text the pipeline built from, and the tree it produced. Deliberately no
 /// tool name is offered to the *check* — spec §1's no-per-tool-logic rule is
@@ -637,1787 +651,6 @@ pub fn find(name: &str) -> anyhow::Result<Box<dyn Detector>> {
                     .join(", ")
             )
         })
-}
-
-// ----------------------------------------------------------------------
-// The detectors themselves
-// ----------------------------------------------------------------------
-
-/// `verbatim-fallback`: help text was captured, and the grammar produced no
-/// structure from it whatsoever.
-///
-/// The check is the tree's own shape rather than anything about the text:
-/// the root carries unparsed lines and has no flags, no subcommands and no
-/// positionals anywhere. That is exactly the state the verbatim tier leaves
-/// behind, and it is why this detector is the harness's proving case — the
-/// condition is unambiguous, so a disagreement with a human label is a fact
-/// about the label rather than about a heuristic's threshold.
-struct VerbatimFallback;
-
-impl Detector for VerbatimFallback {
-    fn name(&self) -> &'static str {
-        "verbatim-fallback"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("verbatim-fallback")
-    }
-    fn describes(&self) -> &'static str {
-        "the root has unparsed lines and no flags, subcommands or positionals anywhere — help \
-         text came back and the grammar made nothing of it"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        let root = evidence.root;
-        if root.unparsed.is_empty() {
-            return Vec::new();
-        }
-        if !tree_is_structureless(root) {
-            return Vec::new();
-        }
-        vec![format!(
-            "{} unparsed line(s) and no extracted structure at all",
-            root.unparsed.len()
-        )]
-    }
-}
-
-/// True when `node` and everything below it carries no flag, no positional
-/// and no child.
-fn tree_is_structureless(node: &CommandNode) -> bool {
-    node.flags().next().is_none()
-        && node.positionals().next().is_none()
-        && node.subcommands.iter().all(tree_is_structureless)
-        && node.subcommands.is_empty()
-}
-
-/// `unparsed-positional`, narrowed to the one shape that can be asserted
-/// without a threshold: argparse prints a literal `positional arguments:`
-/// heading and lists its operands under it, so a tool whose raw text has
-/// that heading and whose root has zero positionals has demonstrably lost
-/// them.
-///
-/// **Deliberately narrower than the family.** `ping4`'s
-/// `<destination DNS name or IP address>` and `vim.basic`'s operands are
-/// the same family and this detector cannot see them; a broader rule over
-/// arbitrary usage lines is exactly where a false-positive rate would come
-/// from. Calibration reports the misses as misses rather than letting a
-/// narrow rule look like a complete one.
-struct UnparsedArgparsePositional;
-
-impl Detector for UnparsedArgparsePositional {
-    fn name(&self) -> &'static str {
-        "unparsed-argparse-positional"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("unparsed-positional")
-    }
-    fn describes(&self) -> &'static str {
-        "raw help has an argparse `positional arguments:` heading with at least one entry under \
-         it, and the extracted root has no positionals"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        if evidence.root.positionals().next().is_some() {
-            return Vec::new();
-        }
-        let listed = argparse_positional_names(evidence.raw);
-        if listed.is_empty() {
-            return Vec::new();
-        }
-        vec![format!(
-            "raw help lists positional argument(s) {} under an argparse heading; the tree has none",
-            listed.join(", ")
-        )]
-    }
-
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        argparse_positional_self_checks()
-    }
-}
-
-/// `uobjnew`'s real bytes (`corpus/rubyobjnew-bpfcc/audit-seed2/help.txt`),
-/// the anchor case for both directions of this detector's self-evidence.
-const UOBJNEW_HELP: &str = "\
-usage: uobjnew [-h] [-l {c,java,ruby,tcl}] [-C TOP_COUNT] [-S TOP_SIZE] [-v] pid [interval]
-
-Summarize object allocations in high-level languages.
-
-positional arguments:
-  pid                   process id to attach to
-  interval              print every specified number of seconds
-
-options:
-  -h, --help            show this help message and exit
-  -v, --verbose         verbose mode: print the BPF program (for debugging purposes)
-";
-
-/// The same heading carrying an `add_subparsers()` group instead of plain
-/// operands — argparse's own rendering, and the nearest false positive this
-/// detector has.
-const ARGPARSE_SUBPARSER_HELP: &str = "\
-usage: widget [-h] {init,build,run} ...
-
-positional arguments:
-  {init,build,run}
-    init            Initialize a new widget
-    build           Build the widget
-    run             Run the widget
-
-options:
-  -h, --help      show this help message and exit
-";
-
-/// An argparse tool with no dashless argument at all: the heading never
-/// appears, and zero root positionals is simply the truth about it.
-const ARGPARSE_FLAGS_ONLY_HELP: &str = "\
-usage: widget [-h] [-v]
-
-options:
-  -h, --help     show this help message and exit
-  -v, --verbose  chatty
-";
-
-/// A block holding both kinds at once — a real operand *and* a subparser
-/// group — which is what stops the subparser rule below from being written
-/// as "any block containing a `{...}` entry is not about positionals".
-const ARGPARSE_MIXED_HELP: &str = "\
-usage: widget [-h] path {init,build} ...
-
-positional arguments:
-  path            the file to process
-  {init,build}
-    init          Initialize a new widget
-    build         Build the widget
-
-options:
-  -h, --help      show this help message and exit
-";
-
-fn argparse_positional_node(name: &str, positionals: &[&str], subcommands: &[&str]) -> CommandNode {
-    let mut root = CommandNode::new(name, Provenance::single(Source::HelpText));
-    root.set_positionals(
-        positionals
-            .iter()
-            .map(|p| {
-                let mut positional = Entity::positional(*p, Provenance::single(Source::HelpText));
-                positional.required = true;
-                positional
-            })
-            .collect(),
-    );
-    root.subcommands = subcommands
-        .iter()
-        .map(|s| CommandNode::new(*s, Provenance::single(Source::HelpText)))
-        .collect();
-    root
-}
-
-/// The hand-built cases this detector is judged on now that the labelled
-/// set has nothing left to confirm (spec §13.1e).
-///
-/// The two `Silent` cases carrying the heading are the ones that do the
-/// work. Without them, "fires on a block of names when the tree has no
-/// positionals" is satisfied by a rule that fires on *every* argparse tool
-/// with a subcommand group, which is the false-alarm class this project
-/// refuses — and the fleet-wide zero this detector will now read would be
-/// unfalsifiable either way.
-fn argparse_positional_self_checks() -> Vec<SelfCheck> {
-    vec![
-        SelfCheck {
-            name: "uobjnew's two operands, dropped",
-            why: "the defect itself: `pid` and `interval` are listed under argparse's own \
-                  heading and the tree has no positionals at all",
-            expect: Expect::Fires(1),
-            raw: UOBJNEW_HELP.to_string(),
-            root: argparse_positional_node("rubyobjnew-bpfcc", &[], &[]),
-        },
-        SelfCheck {
-            name: "uobjnew's two operands, recovered",
-            why: "the other half of the fleet-count-of-zero question: after the fix the same \
-                  bytes must go silent because the operands are in the tree, not because the \
-                  rule stopped working",
-            expect: Expect::Silent,
-            raw: UOBJNEW_HELP.to_string(),
-            root: argparse_positional_node("rubyobjnew-bpfcc", &["pid", "interval"], &[]),
-        },
-        SelfCheck {
-            name: "an add_subparsers() group under the same heading",
-            why: "the nearest real false positive: argparse prints a subcommand list under \
-                  `positional arguments:` too, and a correctly parsed one legitimately leaves \
-                  the root with zero positionals — firing here would report a complete parse as \
-                  a lost operand",
-            expect: Expect::Silent,
-            raw: ARGPARSE_SUBPARSER_HELP.to_string(),
-            root: argparse_positional_node("widget", &[], &["init", "build", "run"]),
-        },
-        SelfCheck {
-            name: "an argparse tool with no operands at all",
-            why: "zero root positionals is the truth for most argparse tools; the heading's \
-                  absence, not the count, is what has to keep this silent",
-            expect: Expect::Silent,
-            raw: ARGPARSE_FLAGS_ONLY_HELP.to_string(),
-            root: argparse_positional_node("widget", &[], &[]),
-        },
-        SelfCheck {
-            name: "a real operand sharing the block with a subparser group",
-            why: "the subparser rule must skip the `{...}` entry and its deeper rows, not the \
-                  whole block — `path` is still a lost operand here, and a rule that discarded \
-                  the block wholesale would report this tool as clean",
-            expect: Expect::Fires(1),
-            raw: ARGPARSE_MIXED_HELP.to_string(),
-            root: argparse_positional_node("widget", &[], &["init", "build"]),
-        },
-    ]
-}
-
-/// Names listed under an argparse `positional arguments:` heading: the
-/// first token of each row at the block's own entry indent, until a blank
-/// or non-indented line ends the block.
-///
-/// Only rows at the entry indent: argparse writes two other things one
-/// level deeper — a wrapped description continuation, and an
-/// `add_subparsers()` group's real subcommands under a `{a,b,c}`
-/// pseudo-entry. Counting those would turn every correctly parsed argparse
-/// subcommand tree into a claimed lost positional. The pseudo-entry itself
-/// is dropped by name; a plain operand sharing the block with it still
-/// counts (exclusion is per row, not per block).
-fn argparse_positional_names(raw: &str) -> Vec<String> {
-    let mut lines = raw.lines();
-    for line in lines.by_ref() {
-        if line
-            .trim_end()
-            .eq_ignore_ascii_case("positional arguments:")
-        {
-            break;
-        }
-    }
-    let mut out = Vec::new();
-    let mut entry_indent: Option<usize> = None;
-    for line in lines {
-        if line.trim().is_empty() {
-            break;
-        }
-        if !line.starts_with(' ') && !line.starts_with('\t') {
-            break;
-        }
-        let Some(word) = line.split_whitespace().next() else {
-            break;
-        };
-        // An option row under this heading means the block already ended and
-        // the layout is not the one this rule understands; stop rather than
-        // report a flag as a positional.
-        if word.starts_with('-') {
-            break;
-        }
-        let indent = line.len() - line.trim_start().len();
-        let base = *entry_indent.get_or_insert(indent);
-        if indent > base {
-            continue;
-        }
-        if word.starts_with('{') {
-            continue;
-        }
-        out.push(word.to_string());
-    }
-    out
-}
-
-/// The bundled-short-flag collapse (`crate::bundling`): a synopsis bundle
-/// of boolean short flags parsed as one flag swallowing the rest as a
-/// value. Zero `crate::existence` fabrications on a collapsed `-2CDlNuVv`
-/// is not a claim of a correct parse — the collapsed token attests cleanly
-/// while the parse destroys seven flags.
-///
-/// Shares a structural fingerprint (`short && !long && value_name`) with
-/// `single-dash-long` and `repeated-char-flag` (all under `k1 = true` in
-/// the labelled set); `crate::bundling` discriminates on what the
-/// swallowed text is, not on structure.
-pub(crate) struct BundledShortFlag;
-
-/// `bundled-short-flag`'s declared exclusions. The one entry is the shape
-/// of every future one: a witness token, the constant it falls below, and
-/// arithmetic that has to agree — see [`Exclusion`].
-const BUNDLED_SHORT_FLAG_EXCLUSIONS: &[Exclusion] = &[Exclusion {
-    tool: "ssh-keygen",
-    ground: Ground::BelowMemberThreshold {
-        cluster: crate::bundling::SSH_KEYGEN_CLUSTER,
-        constant: "bundling::MIN_BUNDLED_MEMBERS",
-        threshold: crate::bundling::MIN_BUNDLED_MEMBERS,
-    },
-    note: "a real collapse this detector knowingly does not claim, not an oversight: at one \
-           swallowed member the shape is genuinely ambiguous, and the fleet scan found the \
-           one-member population is about half correct parses (`xxd -ps`, `which -as`, \
-           `sg_map -st`, `mandoc -ac`)",
-}];
-
-impl Detector for BundledShortFlag {
-    fn name(&self) -> &'static str {
-        "bundled-short-flag"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("bundled-short-flag")
-    }
-    fn describes(&self) -> &'static str {
-        "a synopsis bundle of boolean short flags (`[-abcXYZ]`) parsed as one flag carrying the \
-         rest as a required value, destroying every other member"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::bundling::detect(evidence.raw, evidence.root)
-            .collapses
-            .iter()
-            .map(|c| {
-                format!(
-                    "{:?} at {:?} swallows {} member(s) of the cluster {:?}",
-                    c.spelling, c.path, c.destroyed, c.cluster
-                )
-            })
-            .collect()
-    }
-    fn scope(&self) -> Scope {
-        Scope {
-            claim: "synopsis-sourced short-flag clusters with 2 or more swallowed members \
-                    (`bundling::MIN_BUNDLED_MEMBERS`); a single swallowed member is deliberately \
-                    excluded because the fleet scan found it genuinely ambiguous — see this \
-                    detector's own module doc comment for the measured counter-examples \
-                    (`xxd -ps`, `which -as`, `sg_map -st`, `mandoc -ac`) that a looser threshold \
-                    would false-positive on",
-            known_exclusions: BUNDLED_SHORT_FLAG_EXCLUSIONS,
-        }
-    }
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        crate::bundling::self_checks()
-    }
-}
-
-/// `brace-alternation-flag` (`crate::alternation`), the fourth oracle and
-/// second family detector. Its `hits` calls the same
-/// `help_text::parse_flag_alternation` the grammar calls, rather than
-/// restating the rule — imports the grammar's own rule rather than
-/// risking the drift `crate::misattribution`'s hand-copied `pick_stream`
-/// cost 200 of 656 fabrications.
-///
-/// Fleet count is reported, not ratcheted at zero: `btrfs`'s `btrfs
-/// device scan [-d|--all-devices] <device>` reads the alternation
-/// correctly, but the flags belong to a subcommand node
-/// `unparsed-subcommand` prevents from existing, and `-d` is not a root
-/// flag — reaching further would assert something false. `btrfs` is the
-/// same tool [`UnparsedCommandTable`] excludes as its shape C; whichever
-/// fix lands first zeroes the other's residual for free.
-///
-/// Declared scope names no excluded tool (honestly — both labelled
-/// members, `cache_restore` and `eqn`, are inside the claim), but does
-/// carry the two shapes this detector knowingly does not reach.
-pub(crate) struct BraceAlternationFlag;
-
-impl Detector for BraceAlternationFlag {
-    fn name(&self) -> &'static str {
-        "brace-alternation-flag"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("brace-alternation-flag")
-    }
-    fn describes(&self) -> &'static str {
-        "a delimited alternation of bare flag spellings (`{-i|--input}`, `[[-c|-C] cmd]`) whose \
-         members reach no flag in the tree, or whose surviving member kept the group's own \
-         punctuation as its value"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::alternation::detect(evidence.raw, evidence.root)
-            .findings
-            .iter()
-            .map(|f| format!("{} at {:?}: {}", f.group, f.path, f.detail))
-            .collect()
-    }
-    fn scope(&self) -> Scope {
-        Scope {
-            claim: "delimited groups offering at least 2 bare flag spellings \
-                    (`alternation::MIN_ALTERNATIVES`); a one-member group is an ordinary \
-                    bracketed optional flag the synopsis path already reads correctly, and an \
-                    alternation whose members carry their own values (`sg_sanitize`'s \
-                    `--count=OC|-c OC`) is the value-name-mangled family — genuinely ambiguous \
-                    about whether one value or two are meant, so neither this detector nor the \
-                    grammar claims it",
-            known_exclusions: &[],
-        }
-    }
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        crate::alternation::self_checks()
-    }
-}
-
-pub(crate) struct UnparsedCommandTable;
-
-/// `unparsed-command-table`'s declared exclusions: the three
-/// `unparsed-subcommand` tools that turned out to write their subcommand
-/// lists in entirely different grammars (see `crate::commandtable`'s module
-/// doc comment for the four-shape breakdown). Each cites a real line from
-/// the tool's own capture, and [`Ground::UnreadableEntryShape`] runs this
-/// detector's row parser over it, so none of the three can be excluded by
-/// assertion.
-const UNPARSED_COMMAND_TABLE_EXCLUSIONS: &[Exclusion] = &[
-    Exclusion {
-        tool: "apt-ftparchive",
-        ground: Ground::UnreadableEntryShape {
-            entry: crate::commandtable::APT_FTPARCHIVE_ENTRY,
-            grammar: crate::commandtable::ENTRY_GRAMMAR,
-        },
-        note: "shape B: the `Commands:` label carries the first entry on its own line and the \
-               rest hang under it by alignment, with no dash column anywhere — a different \
-               recovery problem that shares only the audit label",
-    },
-    Exclusion {
-        tool: "btrfs",
-        ground: Ground::UnreadableEntryShape {
-            entry: crate::commandtable::BTRFS_ENTRY,
-            grammar: crate::commandtable::ENTRY_GRAMMAR,
-        },
-        note: "shape C: no command heading exists at all; the names are recoverable only by \
-               stripping the repeated program name off a catalogue of full usage lines, two \
-               levels deep (`btrfs balance start`)",
-    },
-    Exclusion {
-        tool: "ip",
-        ground: Ground::UnreadableEntryShape {
-            entry: crate::commandtable::IP_ENTRY,
-            grammar: crate::commandtable::ENTRY_GRAMMAR,
-        },
-        note: "shape D: the objects are a brace-delimited, pipe-separated alternation set bound \
-               to a metavariable used in the usage line — and this tool's own corpus contract \
-               tests flags (`-V`, `-s`, `-d`), not subcommands, which is its own evidence that \
-               the `unparsed-subcommand` label is not what is chiefly wrong with it",
-    },
-];
-
-impl Detector for UnparsedCommandTable {
-    fn name(&self) -> &'static str {
-        "unparsed-command-table"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("unparsed-subcommand")
-    }
-    fn describes(&self) -> &'static str {
-        "a dash-separated `<name>  - <description>` command table under a `commands:` heading \
-         whose every name is absent from the tree, because the usage-block scanner joined the \
-         heading and all its rows into one synopsis string"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::commandtable::detect(evidence.raw, evidence.root)
-            .missing
-            .iter()
-            .map(|t| {
-                format!(
-                    "{:?} offers {} command(s) {:?}, none of which reached the tree",
-                    t.heading,
-                    t.names.len(),
-                    t.names
-                )
-            })
-            .collect()
-    }
-    fn scope(&self) -> Scope {
-        Scope {
-            claim: "shape A only — a `commands:` heading followed by two or more indented \
-                    `<name>  - <description>` rows (`commandtable::MIN_TABLE_ENTRIES`), with \
-                    none of the names in the tree. The `unparsed-subcommand` label covers four \
-                    unrelated grammars and this detector deliberately reads one of them; the \
-                    other three are named below with a witness line each, because one detector \
-                    loose enough to span all four would be worthless in each",
-            known_exclusions: UNPARSED_COMMAND_TABLE_EXCLUSIONS,
-        }
-    }
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        crate::commandtable::self_checks()
-    }
-}
-
-/// The dropped-alias defect (`crate::dropped_alias`): a flag documented
-/// with both a short and a long spelling reaching the tree with only one
-/// of them, because a value spec interrupted its alias list. Like
-/// [`BundledShortFlag`], the two anti-fabrication oracles are structurally
-/// blind to it — nothing here is invented, what's wrong is what's absent.
-///
-/// Its risk runs the opposite way to every other detector's: over-firing
-/// argues for a fix that would merge two genuinely different flags, and a
-/// fabricated alias is strictly worse than a dropped one. Two of the
-/// seven labelled tools are declared out of scope with structural
-/// grounds rather than reached for.
-pub(crate) struct DroppedAliasDetector;
-
-/// `dropped-alias`'s declared exclusions: the two labelled tools whose
-/// dropped spelling is not separated from its partner by a value spec at
-/// all, each with the witness its [`Ground`] is computed from.
-const DROPPED_ALIAS_EXCLUSIONS: &[Exclusion] = &[
-    Exclusion {
-        tool: "eqn",
-        ground: Ground::InsideAlternationGroup {
-            token: crate::dropped_alias::EQN_VERSION_GROUP,
-            family: "brace-alternation-flag",
-        },
-        note: "A MISLABEL IN THE MANIFEST, reported rather than amended here: `eqn` carries \
-               `dropped-alias` and its shape is brace alternation, which the manifest already \
-               has a family for. The two are not the same defect — a brace group loses the \
-               spelling because the tokenizer never opens the group, not because a value spec \
-               interrupted an alias list — and the same alternation rule closes `cache_restore`'s \
-               `{-i|--input} <file>` and `xfs_io`'s `[[-c|-C] cmd]...` too, neither of which has \
-               a value spec in the way. Claiming it here would let this detector's fleet count \
-               stand in for a fix it did not make, and makes this family look one tool bigger \
-               than it is",
-    },
-    Exclusion {
-        tool: "jdeprscan",
-        ground: Ground::AcrossDescriptionColumn {
-            row: crate::dropped_alias::JDEPRSCAN_LIST_ROW,
-            constant: "help_text::MIN_COLUMN_GAP_SPACES",
-            column_gap: mandible_extract::help_text::MIN_COLUMN_GAP_SPACES,
-        },
-        note: "two shapes in one tool, neither an interrupted alias list: `-l    --list` puts its \
-               long form past the description column — since recovered, by the aligned-spelling- \
-               column split (`help_text::sections::spelling_run`), so this half of the ground is \
-               now historical and the row is kept here because the ground is still *measured* \
-               from it — and `-? -h --help` names a second short that this module's `short()` \
-               accessor has no way to surface (one `Option<char>`), exactly as `-A, --catenate, \
-               --concatenate` names a second long its `long()` accessor cannot surface either; \
-               that second half is what still excludes the tool",
-    },
-];
-
-impl Detector for DroppedAliasDetector {
-    fn name(&self) -> &'static str {
-        "dropped-alias"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("dropped-alias")
-    }
-    fn describes(&self) -> &'static str {
-        "a flag whose value spec interrupted its own alias list: the tool documents `-p PID, \
-         --pid PID` (or `--count=OC|-c OC`) and the tree carries only one of the two spellings"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::dropped_alias::detect(evidence.raw, evidence.root)
-            .drops
-            .iter()
-            .map(|d| {
-                format!(
-                    "{:?} at {:?} keeps {:?} and drops {:?}, documented together as {:?}",
-                    d.kept, d.path, d.kept, d.dropped, d.witness
-                )
-            })
-            .collect()
-    }
-    fn scope(&self) -> Scope {
-        Scope {
-            claim: "alias pairs a VALUE SPEC came between — `-p PID, --pid PID`, \
-                    `--count=OC|-c OC` — where the separator sits at the stored placeholder's own \
-                    boundary and a whole flag spelling follows it. Pairs separated by anything \
-                    else are deliberately not claimed: a wide space run is the description column \
-                    (`jdeprscan`), a brace group is its own labelled family (`eqn`), and a second \
-                    short or a second long has no accessor to reach at all (`Entity::short`/ \
-                    `Entity::long` each surface one). \
-                    Narrow on purpose — the loose rule this replaces would merge two genuinely \
-                    different flags, and a fabricated alias is worse than a dropped one",
-            known_exclusions: DROPPED_ALIAS_EXCLUSIONS,
-        }
-    }
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        crate::dropped_alias::self_checks()
-    }
-}
-
-/// The existence oracle (`crate::existence`), registered so the harness's
-/// answer for an uncalibratable detector is exercised by a real one rather
-/// than only by a test double.
-///
-/// Its `family()` is `None` and that is a finding, not an omission: across
-/// 94 human verdicts, **not one reviewer reported a fabricated subcommand
-/// or flag spelling**. The defect [M-10] shipped — `tar`'s 39 invented
-/// nodes — has no representative in the labelled set, so this set cannot
-/// confirm or refute this oracle at all.
-struct ExistenceOracle;
-
-impl Detector for ExistenceOracle {
-    fn name(&self) -> &'static str {
-        "existence"
-    }
-    fn family(&self) -> Option<&'static str> {
-        None
-    }
-    fn describes(&self) -> &'static str {
-        "a help-text-sourced subcommand name or flag spelling that does not occur in the tool's \
-         own raw text (spec §13.1)"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::existence::detect(evidence.raw, evidence.root)
-            .fabrications
-            .iter()
-            .map(|f| {
-                format!(
-                    "{:?} at {:?} does not occur in the raw text",
-                    f.name, f.path
-                )
-            })
-            .collect()
-    }
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        existence_self_checks()
-    }
-}
-
-/// `tar --help`'s real first line, byte-exact. Both halves of the operand
-/// rule live in this one string: `OPTION` is the slot that names tar's own
-/// flag list, `FILE` is the operand a user passes, and the two are written
-/// in identical notation one slot apart.
-const TAR_SYNOPSIS: &str = "Usage: tar [OPTION...] [FILE]...\n";
-
-/// `uobjnew`'s real argparse synopsis, byte-exact — the counter-shape: a
-/// synopsis that spells its own flags out, where every slot is an operand.
-const UOBJNEW_SYNOPSIS: &str =
-    "usage: uobjnew [-h] [-l {c,java,ruby,tcl}] [-C TOP_COUNT] [-v] pid [interval]\n";
-
-/// `gh --help`'s real shape, byte-exact: a bare `USAGE` heading (no colon,
-/// so it is never a labelled `usage:` marker) followed by a line that opens
-/// with the tool's own name and reads as usage grammar with no marker
-/// anywhere — the **unlabelled** synopsis this task's own fix taught the
-/// oracle to enter. Its trailing `[flags]` is also the anchor case for the
-/// vocabulary fallback in `existence::option_list_slot`: the real flag
-/// stand-in sits *last* here, not first, which is exactly the shape that
-/// would have tripped the position rule into also excluding the genuine
-/// leading operand `command` if the two rules ran unconditionally instead
-/// of one gating the other.
-const GH_UNLABELLED_SYNOPSIS: &str = "USAGE\n  gh <command> <subcommand> [flags]\n";
-
-/// `nfsidmap --help`'s real shape, byte-exact: the C `fprintf(stderr, "%s:
-/// Usage: ...", argv[0])` idiom, which repeats the tool's own name twice on
-/// one line (once as the `fprintf` prefix, once again as the invocation's
-/// own program name) with no ordinary `usage:`-prefixed line anywhere else
-/// in the document.
-const NFSIDMAP_NAME_PREFIXED_SYNOPSIS: &str = "nfsidmap: Usage: nfsidmap [-vh] [-c || -d] path\n";
-
-/// `vgextend --help`'s real shape, byte-exact: LVM's own bare-own-name
-/// convention (the whole `vg*`/`lv*`/`pv*` family shares it) — no docopt
-/// notation anywhere on the invocation line itself, only on the line right
-/// after it. The entire family (29 tools) was newly flagged on this task's
-/// own before/after sweep until this shape was added, every one of them
-/// false.
-const VGEXTEND_BARE_OWN_NAME_SYNOPSIS: &str = concat!(
-    "  vgextend - Add physical volumes to a volume group\n",
-    "\n",
-    "  vgextend VG PV ...\n",
-    "\t[ -A|--autobackup y|n ]\n",
-    "\t[ COMMON_OPTIONS ]\n",
-);
-
-/// A root carrying `positionals` with the given names, all help-text-sourced.
-fn positional_node(name: &str, positionals: &[&str]) -> CommandNode {
-    let mut root = CommandNode::new(name, Provenance::single(Source::HelpText));
-    root.set_positionals(
-        positionals
-            .iter()
-            .map(|p| Entity::positional(*p, Provenance::single(Source::HelpText)))
-            .collect(),
-    );
-    root
-}
-
-/// The existence oracle's own hand-built cases (spec §13.1e).
-///
-/// The oracle is not calibratable against the labelled set — no reviewer in
-/// the seed-2 audit reported a fabricated name — so these cases are the
-/// *only* runtime evidence that it still works, and the operand half needs
-/// them more than the other two: the grammar fix that removed 15 fabricated
-/// operands from the fleet also removed every live example of them, so from
-/// the fleet count alone "zero because the fabrications are gone" and "zero
-/// because nobody ever wired the positional check up" are the same number.
-///
-/// The `Silent` cases are what make the `Fires` ones mean anything. Each is
-/// a real tool whose operand sits in the same notation, one slot over, from
-/// a placeholder — a rule that fired on shape alone would report all of them.
-fn existence_self_checks() -> Vec<SelfCheck> {
-    vec![
-        SelfCheck {
-            name: "tar's OPTION operand",
-            why: "the defect itself, from the tool that shipped it: `OPTION` is lifted out of \
-                  `[OPTION...]`, which names tar's own flag list and is not an argument anyone \
-                  passes",
-            expect: Expect::Fires(1),
-            raw: TAR_SYNOPSIS.to_string(),
-            root: positional_node("tar", &["OPTION", "FILE"]),
-        },
-        SelfCheck {
-            name: "tar's FILE operand alone",
-            why: "the other half of the fleet-count-of-zero question: after the grammar fix the \
-                  same bytes must go silent because the fabricated operand is gone, not because \
-                  the rule stopped looking",
-            expect: Expect::Silent,
-            raw: TAR_SYNOPSIS.to_string(),
-            root: positional_node("tar", &["FILE"]),
-        },
-        SelfCheck {
-            name: "uobjnew's two real operands",
-            why: "the nearest real false positive: a synopsis that writes its own flags needs no \
-                  stand-in for them, so `pid` is an operand in exactly the slot `OPTION` occupies \
-                  in tar's. These two were *recovered* by the same commit that removed the 15; \
-                  reporting them would undo it",
-            expect: Expect::Silent,
-            raw: UOBJNEW_SYNOPSIS.to_string(),
-            root: positional_node("uobjnew", &["pid", "interval"]),
-        },
-        SelfCheck {
-            name: "an operand named nowhere",
-            why: "the base claim underneath the position rule — an operand the document does not \
-                  contain at all is still reported, whatever the synopsis's shape",
-            expect: Expect::Fires(1),
-            raw: TAR_SYNOPSIS.to_string(),
-            root: positional_node("tar", &["TELEPORT"]),
-        },
-        SelfCheck {
-            name: "gh's two real operands from its unlabelled synopsis",
-            why: "this task's own worked example: `command` and `subcommand` are real, occur \
-                  literally in gh's own output, and sit one slot from `flags` — gh's own flag-list \
-                  stand-in, written *last* rather than first. Before this fix the line was entirely \
-                  invisible to the oracle (no `usage:` marker anywhere on it), so both were reported \
-                  as invented; after it, both must go silent and `flags` must not itself become an \
-                  attested operand hiding a genuine fabrication spelled the same way",
-            expect: Expect::Silent,
-            raw: GH_UNLABELLED_SYNOPSIS.to_string(),
-            root: positional_node("gh", &["command", "subcommand"]),
-        },
-        SelfCheck {
-            name: "an operand named nowhere beside gh's real unlabelled synopsis",
-            why: "the same base claim as tar's TELEPORT case, replayed against the new unlabelled \
-                  entry point: recognizing gh's synopsis line must not turn into a blanket amnesty \
-                  for anything claimed beside it",
-            expect: Expect::Fires(1),
-            raw: GH_UNLABELLED_SYNOPSIS.to_string(),
-            root: positional_node("gh", &["command", "teleport"]),
-        },
-        SelfCheck {
-            name: "nfsidmap's real operand from the name-prefixed usage idiom",
-            why: "the C `\"%s: Usage: ...\"` convention (`nfsidmap: Usage: nfsidmap ...`), the \
-                  other synopsis-entry shape this task's fix adds: the tool's own name occurs \
-                  twice on one line, and `path` — the real operand — must still attest correctly \
-                  once both copies are accounted for",
-            expect: Expect::Silent,
-            raw: NFSIDMAP_NAME_PREFIXED_SYNOPSIS.to_string(),
-            root: positional_node("nfsidmap", &["path"]),
-        },
-        SelfCheck {
-            name: "vgextend's two real operands from its bare own-name line",
-            why: "LVM's own convention: the invocation line carries no docopt notation at all, so \
-                  the unlabelled shape above can never find it — only the next physical line's \
-                  bracketed flag row proves it is usage grammar. Newly flagged the whole `vg*`/ \
-                  `lv*`/`pv*` family (29 tools) on this task's own before/after sweep until this \
-                  shape was added, every one of them false",
-            expect: Expect::Silent,
-            raw: VGEXTEND_BARE_OWN_NAME_SYNOPSIS.to_string(),
-            root: positional_node("vgextend", &["VG", "PV"]),
-        },
-        SelfCheck {
-            name: "an operand named nowhere beside vgextend's real bare own-name line",
-            why: "the same base claim replayed a third time: recognizing this entry shape must not \
-                  turn into a blanket amnesty for anything claimed beside it",
-            expect: Expect::Fires(1),
-            raw: VGEXTEND_BARE_OWN_NAME_SYNOPSIS.to_string(),
-            root: positional_node("vgextend", &["VG", "TELEPORT"]),
-        },
-    ]
-}
-
-/// The misattribution oracle (`crate::misattribution`), registered on the
-/// same terms as [`ExistenceOracle`]. Its shape — a flag's description
-/// belonging to a different flag — is adjacent to `section-header-bleed`
-/// and to `missing-flag-description`, but adjacency is not identity, and
-/// mapping it onto either would manufacture a matrix from a correspondence
-/// nobody verified.
-struct MisattributionOracle;
-
-impl Detector for MisattributionOracle {
-    fn name(&self) -> &'static str {
-        "misattribution"
-    }
-    fn family(&self) -> Option<&'static str> {
-        None
-    }
-    fn describes(&self) -> &'static str {
-        "a flag description containing another flag's spelling, attested at a column-aligned \
-         position elsewhere in the same raw text (spec §13.1)"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::misattribution::detect(evidence.raw, evidence.root)
-            .suspects
-            .iter()
-            .map(|s| format!("{:?} at {:?}", s.flag, s.path))
-            .collect()
-    }
-}
-
-// ----------------------------------------------------------------------
-// Calibration
-// ----------------------------------------------------------------------
-
-/// One audited tool as calibration sees it.
-pub struct Case {
-    pub tool: String,
-    /// The tool's derived family labels, empty for a `correct` verdict and
-    /// for an unclassified defect alike — [`Case::expected`] is what
-    /// distinguishes them.
-    pub families: Vec<String>,
-    /// `true` when a human judged this tool's parse defective
-    /// (`wrong`/`incomplete`), `false` when they judged it `correct`.
-    /// `skip` never becomes a `Case` at all.
-    pub judged_defect: bool,
-    /// The replayed fixture, or `None` when this tool has none — the
-    /// not-evaluable population.
-    pub evidence: Option<ReplayedCase>,
-}
-
-/// A case's replayed bytes and tree, kept out of [`Case`] itself so a tool
-/// with no fixture is representable without a fake tree.
-pub struct ReplayedCase {
-    pub raw: String,
-    pub root: CommandNode,
-}
-
-impl Case {
-    /// Whether this tool is *expected* to fire for `family`: it must be a
-    /// human-judged defect **and** carry that family label. A judged defect
-    /// with no label (unclassified) is expected-silent, deliberately — an
-    /// unclassified tool cannot be counted as a miss for a family nobody
-    /// established it belongs to.
-    fn expected(&self, family: &str) -> bool {
-        self.judged_defect && self.families.iter().any(|f| f == family)
-    }
-}
-
-/// One detector's confusion matrix against the labelled set, plus the two
-/// populations that are not in it.
-pub struct Calibration {
-    pub detector: &'static str,
-    pub describes: &'static str,
-    /// `None` when the detector generalizes no labelled family — every cell
-    /// below is then empty and [`render`] says why instead of printing a
-    /// matrix of zeroes that would read as a perfect score.
-    pub family: Option<&'static str>,
-    /// The detector's own declared scope ([`Detector::scope`]) — carried on
-    /// the result so [`render`] can print the whole declared exclusion list
-    /// unconditionally, independent of what this particular run found.
-    pub scope: Scope,
-    /// Labelled with the family, and the detector fired. `(tool, reasons)`.
-    pub true_positives: Vec<(String, Vec<String>)>,
-    /// Labelled with the family, and the detector was silent — **and the
-    /// tool is not a declared exclusion of [`Calibration::scope`].** This is
-    /// the cell recall is computed over, i.e. recall *within* the declared
-    /// scope.
-    pub false_negatives: Vec<String>,
-    /// Labelled with the family, the detector was silent, **and the tool is
-    /// named in [`Calibration::scope`]'s declared exclusions.** A miss the
-    /// detector never claimed to catch — counted, named and reasoned
-    /// separately from [`Calibration::false_negatives`] rather than folded
-    /// into recall, but never dropped from the report: [`render`] prints
-    /// every declared exclusion every time, in red, whether or not this run
-    /// found it here.
-    pub out_of_scope_misses: Vec<(String, &'static Exclusion)>,
-    /// Judged `correct` by a human, and the detector fired anyway.
-    pub false_alarms: Vec<(String, Vec<String>)>,
-    /// Judged `correct` by a human, and the detector stayed silent.
-    pub true_negatives: Vec<String>,
-    /// Judged a defect, but of some *other* family (or unclassified), and
-    /// the detector fired. Reported in its own cell rather than folded into
-    /// false alarms: the human already said this tool's parse is wrong, so
-    /// a fire here is a possible mislabel or a genuine second family, not
-    /// evidence that the detector is noisy. Reading it as a false positive
-    /// would understate the detector; reading it as a true positive would
-    /// overstate it. It is neither.
-    pub fires_on_other_defect: Vec<(String, Vec<String>)>,
-    /// Judged, labelled, but with no fixture to replay — evaluated by
-    /// nobody, counted by name.
-    pub not_evaluable: Vec<String>,
-    /// Judged defects carrying no family label at all, across the whole
-    /// manifest. Not a cell of this detector's matrix; printed with it
-    /// because it bounds how complete *any* family's calibration can be.
-    pub unclassified: Vec<String>,
-    /// The detector's own hand-built cases, re-run during this calibration
-    /// ([`run_self_checks`]). Not a cell of the matrix and never mixed into
-    /// one: these are not labelled tools and cannot substitute for them.
-    /// They answer a different question — is the *rule* still alive — which
-    /// is the only question left once the family has been repaired.
-    pub self_checks: Vec<SelfCheckOutcome>,
-}
-
-/// What a calibration run concluded. Three states, not two.
-///
-/// The third exists because the bundled-short-flag family was actually
-/// repaired (spec §13.1e, "a fixed family inverts its own calibration"), and
-/// with only two states the harness had to render a healthy detector in the
-/// vocabulary of failure: `recall 0%`, `DOES NOT PASS`, six tools listed as
-/// misses. Every word true, the impression wrong.
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub enum Verdict {
-    /// Fires on every evaluable labelled tool it claims, and on no tool a
-    /// human judged correct.
-    Passes,
-    /// The family this detector generalizes has been **fixed**, so the
-    /// labelled set — recorded against the pre-fix parser — has nothing
-    /// left to confirm.
-    ///
-    /// **This is a claim with evidence attached, never a suppression.**
-    /// Reaching it requires the detector's own hand-built self-checks to
-    /// still hold ([`self_checks_are_conclusive`]), which is exactly the
-    /// evidence that separates "zero because the bug is gone" from "zero
-    /// because the detector stopped working". Nothing is moved between
-    /// cells to get here: recall still reads 0%, and every missed tool is
-    /// still counted and still printed by name.
-    Repaired,
-    /// Anything else — including a family whose calibration has inverted
-    /// but whose self-checks did *not* hold, which is the genuinely broken
-    /// detector this state must never be confused with.
-    DoesNotPass,
-}
-
-impl Calibration {
-    /// Fired-when-expected over expected, **within the detector's declared
-    /// scope** — [`Calibration::out_of_scope_misses`] is deliberately not in
-    /// either term, because a detector is scored against what it claims,
-    /// not against every shape of the family. `None` when the family has no
-    /// evaluable labelled member in scope.
-    pub fn recall(&self) -> Option<f64> {
-        let expected = self.true_positives.len() + self.false_negatives.len();
-        (expected > 0).then(|| self.true_positives.len() as f64 / expected as f64)
-    }
-
-    /// Silent-when-judged-correct over judged-correct, `None` when no
-    /// evaluable `correct` case exists.
-    pub fn silence(&self) -> Option<f64> {
-        let good = self.true_negatives.len() + self.false_alarms.len();
-        (good > 0).then(|| self.true_negatives.len() as f64 / good as f64)
-    }
-
-    /// Whether this detector's own self-checks are strong enough to stand
-    /// in for a labelled set with nothing left to confirm.
-    pub fn self_checks_are_conclusive(&self) -> bool {
-        self_checks_are_conclusive(&self.self_checks)
-    }
-
-    /// Whether calibration has *inverted*: there were labelled members to
-    /// fire on, and not one of them fires any more.
-    ///
-    /// Necessary for [`Verdict::Repaired`] and nowhere near sufficient — on
-    /// its own this is the exact signature of a detector that has stopped
-    /// working, which is why the verdict also demands the self-checks.
-    pub fn calibration_inverted(&self) -> bool {
-        self.true_positives.is_empty()
-            && (!self.false_negatives.is_empty() || !self.out_of_scope_misses.is_empty())
-    }
-
-    /// The run's conclusion — see [`Verdict`] for why there are three.
-    ///
-    /// [`Verdict::Passes`] is the precondition itself: fires on every
-    /// labelled member within its declared scope, and never on a tool a
-    /// human judged correct. Silence on `correct` tools is required
-    /// absolutely — scope narrows what a detector may be scored on
-    /// *missing*, never what excuses it for *firing wrongly*, so
-    /// [`Calibration::false_alarms`] blocks every verdict but
-    /// [`Verdict::DoesNotPass`]. Recall is required only over what is
-    /// evaluable and in scope: a detector cannot be blamed for a tool with
-    /// no fixture, nor for a tool it never claimed. A named out-of-scope
-    /// miss never blocks a pass, and it never stops being printed either —
-    /// see [`render`].
-    pub fn verdict(&self) -> Verdict {
-        // A false alarm blocks everything, unconditionally and in every
-        // state: firing on a tool a human judged correct is never excused
-        // by a declared scope and never excused by a repaired family. This
-        // project's standing rule, checked first.
-        if self.family.is_none() || !self.false_alarms.is_empty() {
-            return Verdict::DoesNotPass;
-        }
-        if self.false_negatives.is_empty() && !self.true_positives.is_empty() {
-            return Verdict::Passes;
-        }
-        if self.calibration_inverted() && self.self_checks_are_conclusive() {
-            return Verdict::Repaired;
-        }
-        Verdict::DoesNotPass
-    }
-}
-
-/// Load every judged audit entry as a [`Case`], replaying its
-/// `corpus/<tool>/<fixture_version>/` fixture when one exists.
-///
-/// Validates the manifest's family labels first ([`AuditFile::validate_families`]),
-/// so a mistyped family fails here rather than quietly shrinking a cell.
-pub fn load_cases(
-    audit_file: &AuditFile,
-    corpus_root: &Path,
-    fixture_version: &str,
-) -> anyhow::Result<Vec<Case>> {
-    audit_file.validate_families()?;
-    let mut replayed: BTreeMap<String, ReplayedCase> = BTreeMap::new();
-    for fixture in corpus::replay_version(corpus_root, fixture_version)? {
-        if let Some(root) = fixture.root {
-            replayed.insert(
-                fixture.tool,
-                ReplayedCase {
-                    raw: fixture.raw,
-                    root,
-                },
-            );
-        }
-    }
-
-    let mut cases = Vec::new();
-    for entry in &audit_file.entries {
-        let judged_defect = entry.is_judged_defect();
-        if !judged_defect && !entry.is_judged_correct() {
-            continue;
-        }
-        cases.push(Case {
-            tool: entry.tool.clone(),
-            families: entry.families.clone(),
-            judged_defect,
-            evidence: replayed.remove(&entry.tool),
-        });
-    }
-    Ok(cases)
-}
-
-/// Run `detector` over `cases` and sort every one into its cell.
-pub fn calibrate(
-    detector: &dyn Detector,
-    cases: &[Case],
-    unclassified: Vec<String>,
-) -> Calibration {
-    let scope = detector.scope();
-    let mut cal = Calibration {
-        detector: detector.name(),
-        describes: detector.describes(),
-        family: detector.family(),
-        scope,
-        true_positives: Vec::new(),
-        false_negatives: Vec::new(),
-        out_of_scope_misses: Vec::new(),
-        false_alarms: Vec::new(),
-        true_negatives: Vec::new(),
-        fires_on_other_defect: Vec::new(),
-        not_evaluable: Vec::new(),
-        unclassified,
-        // Re-run every time, on the same run that computes the matrix: the
-        // two have to be read together, and a cached or separately-invoked
-        // self-check could be stale exactly when it matters most.
-        self_checks: run_self_checks(detector),
-    };
-    let Some(family) = detector.family() else {
-        return cal;
-    };
-
-    for case in cases {
-        let Some(evidence) = &case.evidence else {
-            // A tool with no fixture is only worth naming when the labels
-            // say it *should* have fired: an unevaluated `correct` tool
-            // contributes nothing either way and would only pad the list.
-            if case.expected(family) {
-                cal.not_evaluable.push(case.tool.clone());
-            }
-            continue;
-        };
-        let hits = detector.hits(&ToolEvidence {
-            raw: &evidence.raw,
-            root: &evidence.root,
-        });
-        match (case.expected(family), case.judged_defect, hits.is_empty()) {
-            (true, _, false) => cal.true_positives.push((case.tool.clone(), hits)),
-            // A labelled miss is only ever *reclassified*, never dropped: a
-            // declared exclusion moves it from false_negatives (which blocks
-            // a pass) to out_of_scope_misses (which does not) — it is still
-            // named, still counted, and render() prints the whole declared
-            // list unconditionally regardless of which cell it lands in.
-            (true, _, true) => {
-                let exclusion = cal
-                    .scope
-                    .known_exclusions
-                    .iter()
-                    .find(|e| e.tool == case.tool);
-                match exclusion {
-                    Some(e) => cal.out_of_scope_misses.push((case.tool.clone(), e)),
-                    None => cal.false_negatives.push(case.tool.clone()),
-                }
-            }
-            (false, true, false) => cal.fires_on_other_defect.push((case.tool.clone(), hits)),
-            (false, true, true) => {}
-            // Scope narrows what a detector may be scored on missing; it
-            // never excuses firing on a tool a human judged correct. A false
-            // alarm reaches this arm unconditionally, declared exclusion or
-            // not.
-            (false, false, false) => cal.false_alarms.push((case.tool.clone(), hits)),
-            (false, false, true) => cal.true_negatives.push(case.tool.clone()),
-        }
-    }
-    cal
-}
-
-// ----------------------------------------------------------------------
-// Rendering
-// ----------------------------------------------------------------------
-
-/// The caveat that must appear above every calibration result, in full,
-/// every time.
-///
-/// Not a footnote and not abbreviated on repeat runs: a number that travels
-/// (into a commit message, an issue, a release note) travels without its
-/// context unless the context is printed next to it. This is the same
-/// discipline `accuracy: unmeasured` enforces on the coverage scoreboard
-/// (spec §13.1b).
-fn caveat(set: &SetSize, unclassified: usize) -> String {
-    let SetSize {
-        sampled,
-        judged,
-        evaluable,
-    } = *set;
-    format!(
-        "CALIBRATION IS AGAINST DERIVED LABELS OVER A BOUNDED SAMPLE — NOT GROUND TRUTH ABOUT \
-         THE FLEET.\n  * the verdicts are human: {judged} judged tools out of {sampled} sampled \
-         in the seed-2 audit (the rest were skipped, and a skip judges nothing either way). The \
-         defect-family labels on them are a MACHINE READING of each reviewer's note plus the \
-         fixture evidence (`families_derived = true`) — a weaker claim than the verdict it sits \
-         on\n  * {sampled} tools is roughly 4% of PATH. Passing here means a detector works on \
-         these tools; it says nothing about whether its fleet-wide count is right\n  * \
-         {evaluable} of the {judged} have a replayable fixture. The rest are evaluated by nobody \
-         and are listed as not-evaluable rather than quietly dropped\n  * {unclassified} judged \
-         defect(s) carry no family label at all, so no family's calibration here can be complete"
-    )
-}
-
-/// How big the labelled set is, in the three counts every calibration
-/// report must state: what the audit sampled, how much of that carries a
-/// judgment at all, and how much of *that* has frozen bytes to replay.
-/// Passed around as one value so a caller cannot print two of the three and
-/// leave a reader to assume the missing one.
-#[derive(Clone, Copy)]
-pub struct SetSize {
-    /// Entries in the manifest, skips included.
-    pub sampled: usize,
-    /// Entries with a `wrong`/`incomplete`/`correct` verdict — the ones a
-    /// detector can be right or wrong about.
-    pub judged: usize,
-    /// Judged entries that also have a replayable fixture.
-    pub evaluable: usize,
-}
-
-/// Render one detector's calibration as plain text.
-pub fn render(cal: &Calibration, set: &SetSize) -> String {
-    let mut s = String::new();
-    s.push_str(&format!("detector: {}\n", cal.detector));
-    s.push_str(&format!("checks:   {}\n", cal.describes));
-    match cal.family {
-        Some(f) => s.push_str(&format!("family:   {f}\n")),
-        None => s.push_str("family:   (none in the labelled set)\n"),
-    }
-    s.push_str(&format!("scope:    {}\n", cal.scope.claim));
-    s.push('\n');
-    s.push_str(&caveat(set, cal.unclassified.len()));
-    s.push_str("\n\n");
-
-    if cal.family.is_none() {
-        s.push_str(
-            "NOT CALIBRATABLE against this set: this detector generalizes no defect family any \
-             reviewer in the seed-2 audit recorded, so these tools can neither confirm nor refute \
-             it. That is a property of the sample, not a defect in the detector — but its \
-             fleet-wide number is not quotable on the strength of anything here.\n",
-        );
-        return s;
-    }
-
-    let row = |label: &str, n: usize| format!("  {label:<46}{n}\n");
-    s.push_str(&row(
-        "fires on labelled-bad   (true positive)",
-        cal.true_positives.len(),
-    ));
-    s.push_str(&row(
-        "silent on labelled-bad  (FALSE NEGATIVE)",
-        cal.false_negatives.len(),
-    ));
-    s.push_str(&format!(
-        "  {RED}{:<46}{}{RESET}\n",
-        "silent on labelled-bad, DECLARED OUT OF SCOPE",
-        cal.out_of_scope_misses.len()
-    ));
-    s.push_str(&row(
-        "silent on labelled-good (true negative)",
-        cal.true_negatives.len(),
-    ));
-    s.push_str(&row(
-        "fires on labelled-good  (FALSE ALARM)",
-        cal.false_alarms.len(),
-    ));
-    s.push_str(&row(
-        "fires on a defect of another family",
-        cal.fires_on_other_defect.len(),
-    ));
-    s.push_str(&row(
-        "labelled-bad with no fixture (not evaluable)",
-        cal.not_evaluable.len(),
-    ));
-    s.push('\n');
-    match cal.recall() {
-        Some(r) => s.push_str(&format!(
-            "  recall over evaluable labelled: {:.0}%\n",
-            r * 100.0
-        )),
-        None => s.push_str("  recall: no evaluable labelled member — nothing demonstrated\n"),
-    }
-    match cal.silence() {
-        Some(v) => s.push_str(&format!(
-            "  silence over judged-correct:    {:.0}%\n",
-            v * 100.0
-        )),
-        None => s.push_str("  silence: no evaluable judged-correct tool\n"),
-    }
-    s.push('\n');
-    s.push_str(&verdict_text(cal));
-
-    // Printed unconditionally, for the same reason as the out-of-scope list
-    // below: it is the evidence the REPAIRED verdict rests on, so it must be
-    // visible on the runs that *don't* reach it too — otherwise the first
-    // time a reader sees this block is the run where it is being used to
-    // excuse a zero.
-    s.push('\n');
-    s.push_str(&render_self_checks(&cal.self_checks));
-
-    // Printed unconditionally — a declared exclusion is a permanent part of
-    // this detector's identity (`Detector::scope`), not data that can
-    // happen to be empty this run. A PASSing verdict above must never be
-    // read as "nothing was missed": if this detector declares any
-    // exclusion, the tool and the reason it doesn't count sit right here,
-    // in red, next to the verdict that would otherwise look complete
-    // without them.
-    s.push_str(&format!(
-        "\n{RED}KNOWN OUT-OF-SCOPE MISSES — declared in code, not measured, never suppressed \
-         ({} declared):{RESET}\n",
-        cal.scope.known_exclusions.len()
-    ));
-    if cal.scope.known_exclusions.is_empty() {
-        s.push_str("  (this detector declares no exclusion — its scope is the full family)\n");
-    } else {
-        for exclusion in cal.scope.known_exclusions {
-            let tool = exclusion.tool;
-            s.push_str(&format!(
-                "  {RED}{tool}{RESET} — {}\n      why out of scope (structural): {}\n      \
-                 note: {}\n",
-                exclusion_status(cal, tool),
-                exclusion.ground.explain(),
-                exclusion.note,
-            ));
-        }
-    }
-
-    s.push_str(&named(
-        "fires on labelled (true positives)",
-        &cal.true_positives,
-    ));
-    s.push_str(&plain(
-        "MISSED labelled tools (false negatives)",
-        &cal.false_negatives,
-    ));
-    s.push_str(&named(
-        "FALSE ALARMS on judged-correct tools",
-        &cal.false_alarms,
-    ));
-    s.push_str(&named(
-        "fires on a defect labelled as another family (neither cell — check by hand)",
-        &cal.fires_on_other_defect,
-    ));
-    s.push_str(&plain(
-        "labelled, no fixture (not evaluable)",
-        &cal.not_evaluable,
-    ));
-    s.push_str(&plain(
-        "silent on judged-correct (true negatives)",
-        &cal.true_negatives,
-    ));
-    s.push_str(&plain(
-        "judged defects with no family label at all (unclassified)",
-        &cal.unclassified,
-    ));
-    s
-}
-
-/// The verdict paragraph — the one place a reader's impression is actually
-/// formed, which is why [`Verdict::Repaired`] has to be able to say so in
-/// its own words rather than being rendered as a failure.
-fn verdict_text(cal: &Calibration) -> String {
-    match cal.verdict() {
-        Verdict::Passes => "VERDICT: PASSES calibration within its declared scope. Fires on \
-                            every evaluable labelled tool it claims and on no \
-                            human-judged-correct one. Its fleet-wide count may be quoted — with \
-                            the caveat above attached to it, AND with the out-of-scope misses \
-                            named immediately below, which a pass never erases.\n"
-            .to_string(),
-        Verdict::Repaired => format!(
-            "VERDICT: REPAIRED — this family was FIXED, and the evidence for saying so is \
-             printed below.\n  Calibration has inverted: {} labelled tool(s) that once fired \
-             are silent, and there is nothing left in the labelled set to confirm against. \
-             Those labels were recorded against the PRE-FIX parser, and spec §13.1e says the \
-             precondition expires for a family on the commit that repairs it.\n  What carries \
-             the weight instead: all {} of this detector's own hand-built self-checks still \
-             hold, including {} case(s) it must fire on and {} it must stay silent on. That is \
-             the evidence — and the ONLY evidence — that separates \"zero because the bug is \
-             gone\" from \"zero because the detector stopped working\"; the fleet number alone \
-             cannot tell them apart.\n  NOTHING IS SUPPRESSED TO REACH THIS VERDICT. Recall \
-             above still reads 0%, every missed tool is still counted in the FALSE NEGATIVE \
-             cell and still named below, and the declared out-of-scope miss is still printed in \
-             red. A repaired family is not a clean bill of health for the fleet: run \
-             `sweep-diff`, which is the instrument that answers whether fixing this broke \
-             anything else.\n",
-            cal.false_negatives.len() + cal.out_of_scope_misses.len(),
-            cal.self_checks.len(),
-            cal.self_checks
-                .iter()
-                .filter(|o| matches!(o.expect, Expect::Fires(_)))
-                .count(),
-            cal.self_checks
-                .iter()
-                .filter(|o| o.expect == Expect::Silent)
-                .count(),
-        ),
-        Verdict::DoesNotPass if cal.calibration_inverted() => format!(
-            "{RED}VERDICT: DOES NOT PASS calibration — and this is the dangerous shape, not the \
-             ordinary one.{RESET}\n  Nothing labelled fires any more, which would be consistent \
-             with the family having been repaired — but this detector's own self-checks DO NOT \
-             back that reading, so a repaired family and a broken detector cannot be told apart \
-             here. See the self-check block below for which case(s) stopped holding; a \
-             detector that no longer fires on the hand-built defective shape is broken, and its \
-             fleet-wide zero means nothing at all.\n"
-        ),
-        Verdict::DoesNotPass => "VERDICT: DOES NOT PASS calibration. Its fleet-wide count is not \
-                                 quotable yet; see the named tools below.\n"
-            .to_string(),
-    }
-}
-
-/// ANSI red, used only for the known-out-of-scope-misses section and for a
-/// self-check that stopped holding: the parts of the report that must be
-/// visually impossible to skim past even in a run that otherwise looks
-/// clean.
-const RED: &str = "\x1b[31m";
-const RESET: &str = "\x1b[0m";
-
-/// What happened to one of a detector's declared exclusions in this
-/// particular calibration run, cross-checked against every cell it could
-/// have landed in rather than assumed to be a miss.
-fn exclusion_status(cal: &Calibration, tool: &str) -> &'static str {
-    if cal.out_of_scope_misses.iter().any(|(t, _)| t == tool) {
-        "MISSED this run, exactly as declared"
-    } else if cal.true_positives.iter().any(|(t, _)| t == tool) {
-        "fired anyway this run — the declared exclusion did not hold; check by hand"
-    } else if cal.not_evaluable.iter().any(|t| t == tool) {
-        "no fixture this run — not evaluable, so this exclusion was not exercised"
-    } else {
-        "not in this labelled set at all — this exclusion was not exercised"
-    }
-}
-
-fn named(title: &str, rows: &[(String, Vec<String>)]) -> String {
-    if rows.is_empty() {
-        return String::new();
-    }
-    let mut s = format!("\n{title}:\n");
-    for (tool, reasons) in rows {
-        s.push_str(&format!("  {tool}\n"));
-        for reason in reasons.iter().take(3) {
-            s.push_str(&format!("      {reason}\n"));
-        }
-        if reasons.len() > 3 {
-            s.push_str(&format!("      ... and {} more\n", reasons.len() - 3));
-        }
-    }
-    s
-}
-
-fn plain(title: &str, rows: &[String]) -> String {
-    if rows.is_empty() {
-        return String::new();
-    }
-    format!("\n{title}:\n  {}\n", rows.join(", "))
-}
-
-// ----------------------------------------------------------------------
-// Commands
-// ----------------------------------------------------------------------
-
-/// `xtask detector list`: every registered detector, its family, and how
-/// many labelled tools that family has.
-pub fn cmd_list(dir: &Path, seed: u64) -> anyhow::Result<()> {
-    let file = audit::load(&audit::verdict_path(dir, seed))?;
-    file.validate_families()?;
-    let mut counts: BTreeMap<&str, usize> = BTreeMap::new();
-    for entry in &file.entries {
-        if entry.is_judged_defect() {
-            for family in &entry.families {
-                *counts.entry(family_static(family)).or_default() += 1;
-            }
-        }
-    }
-
-    println!("registered detectors:\n");
-    for d in registry() {
-        let family = match d.family() {
-            Some(f) => format!(
-                "{f} ({} labelled tool(s))",
-                counts.get(f).copied().unwrap_or(0)
-            ),
-            None => "(none in the labelled set — not calibratable)".to_string(),
-        };
-        println!(
-            "  {}\n      family: {family}\n      checks: {}\n",
-            d.name(),
-            d.describes()
-        );
-    }
-
-    println!("defect families in audit/{seed}.toml (derived labels):");
-    for family in audit::family_names() {
-        println!(
-            "  {:<26} {:>2} labelled  — {}",
-            family,
-            counts.get(family).copied().unwrap_or(0),
-            audit::family_meaning(family).unwrap_or(""),
-        );
-    }
-    let unclassified: Vec<&str> = file.unclassified().map(|e| e.tool.as_str()).collect();
-    println!(
-        "\n  {:<26} {:>2} judged defect(s) carry no family label{}",
-        "(unclassified)",
-        unclassified.len(),
-        if unclassified.is_empty() {
-            String::new()
-        } else {
-            format!(": {}", unclassified.join(", "))
-        }
-    );
-    Ok(())
-}
-
-/// Resolve a family word coming out of the manifest to its `'static`
-/// spelling. Safe by construction: `validate_families` has already rejected
-/// anything outside the set by the time this is reached.
-fn family_static(word: &str) -> &'static str {
-    audit::parse_family(word).unwrap_or("(unrecognized)")
-}
-
-/// `xtask detector calibrate`: the confusion matrix, for one detector or
-/// for all of them.
-pub fn cmd_calibrate(
-    dir: &Path,
-    seed: u64,
-    corpus_root: &Path,
-    fixture_version: &str,
-    detector: Option<&str>,
-) -> anyhow::Result<()> {
-    let file = audit::load(&audit::verdict_path(dir, seed))?;
-    let cases = load_cases(&file, corpus_root, fixture_version)?;
-    let unclassified: Vec<String> = file.unclassified().map(|e| e.tool.clone()).collect();
-    let set = SetSize {
-        sampled: file.entries.len(),
-        judged: cases.len(),
-        evaluable: cases.iter().filter(|c| c.evidence.is_some()).count(),
-    };
-
-    let detectors = match detector {
-        Some(name) => vec![find(name)?],
-        None => registry(),
-    };
-    for d in detectors {
-        let cal = calibrate(d.as_ref(), &cases, unclassified.clone());
-        println!("{}", "=".repeat(76));
-        println!("{}", render(&cal, &set));
-    }
-    Ok(())
-}
-
-// ----------------------------------------------------------------------
-// The ratchet gate
-// ----------------------------------------------------------------------
-
-/// One detector's ratchet-at-zero result: the fleet counts, the self-check
-/// evidence they have to be read against, and every reason the gate refused.
-pub struct RatchetOutcome {
-    pub detector: &'static str,
-    /// What the sweep's scoreboard reported for this detector's family.
-    pub tools: usize,
-    pub destroyed_flags: usize,
-    pub self_checks: Vec<SelfCheckOutcome>,
-    /// Empty when the gate holds. Each entry is one independent reason.
-    pub failures: Vec<String>,
-}
-
-impl RatchetOutcome {
-    pub fn holds(&self) -> bool {
-        self.failures.is_empty()
-    }
-
-    /// The full report, printed whether the gate holds or not — the counts
-    /// are meaningless without the evidence beside them, so they are never
-    /// printed apart.
-    pub fn report(&self) -> String {
-        let mut s = format!(
-            "RATCHET GATE — {} is gated at zero, with evidence.\n  fleet: {} tool(s) with a \
-             collapse, {} real flag(s) destroyed (both must be 0)\n\n",
-            self.detector, self.tools, self.destroyed_flags
-        );
-        s.push_str(&render_self_checks(&self.self_checks));
-        s.push('\n');
-        if self.holds() {
-            s.push_str(
-                "GATE HOLDS: the fleet count is zero AND the detector still fires on its own \
-                 hand-built defective shape while staying silent on the correct parses that \
-                 resemble it. Both halves are required — see `ratchet_at_zero`.\n",
-            );
-        } else {
-            s.push_str(&format!("{RED}GATE FAILS:{RESET}\n"));
-            for failure in &self.failures {
-                s.push_str(&format!("  {RED}*{RESET} {failure}\n"));
-            }
-        }
-        s
-    }
-}
-
-/// Gate `detector`'s fleet-wide count at zero — and refuse to accept that
-/// zero without evidence the detector still works.
-///
-/// A gate on `count == 0` alone is satisfied by deleting the detector, by
-/// `hits()` returning `Vec::new()`, by any refactor that quietly stops the
-/// rule firing — the same "a metric improved by breaking what measures
-/// it" failure spec §13.1b already records twice ([M-10], `%flags_text`).
-/// Spec §13.1e: zero-because-fixed and zero-because-broken are
-/// indistinguishable from the fleet number alone.
-///
-/// So the gate needs both: (1) the detector's self-checks still hold,
-/// conclusively ([`self_checks_are_conclusive`]), and (2) the fleet counts
-/// are zero. Deleting the detector fails (1) and can never reach (2).
-pub fn ratchet_at_zero(
-    detector: &dyn Detector,
-    tools: usize,
-    destroyed_flags: usize,
-) -> RatchetOutcome {
-    let self_checks = run_self_checks(detector);
-    let mut failures = Vec::new();
-
-    // Half 1 first, deliberately: the counts mean nothing until the
-    // instrument that produced them is shown to be alive.
-    if self_checks.is_empty() {
-        failures.push(format!(
-            "{} declares NO self-check, so a fleet count of zero cannot be distinguished from \
-             the detector having been deleted. A gate on the count alone is satisfied by \
-             deleting the detector — that is the whole reason this half exists.",
-            detector.name()
-        ));
-    } else {
-        for outcome in &self_checks {
-            if !outcome.held {
-                failures.push(format!(
-                    "self-check {:?} did not hold: it {} but reported {} hit(s) {:?}. The \
-                     detector no longer behaves as its own evidence says it must, so its \
-                     fleet-wide zero means nothing.",
-                    outcome.name,
-                    match outcome.expect {
-                        Expect::Fires(n) => format!("must fire {n} time(s)"),
-                        Expect::Silent => "must stay silent".to_string(),
-                    },
-                    outcome.hits.len(),
-                    outcome.hits,
-                ));
-            }
-        }
-        if !self_checks
-            .iter()
-            .any(|o| matches!(o.expect, Expect::Fires(_)))
-        {
-            failures.push(format!(
-                "{} declares no self-check it must FIRE on. Without one, nothing here shows the \
-                 rule is still alive.",
-                detector.name()
-            ));
-        }
-        if !self_checks.iter().any(|o| o.expect == Expect::Silent) {
-            failures.push(format!(
-                "{} declares no self-check it must STAY SILENT on. Without one, a detector \
-                 firing indiscriminately would satisfy the must-fire half — and this project's \
-                 standing rule is no false positives over recall.",
-                detector.name()
-            ));
-        }
-    }
-
-    // Half 2: the ratchet itself.
-    if tools != 0 {
-        failures.push(format!(
-            "{tools} tool(s) exhibit this collapse; the ratchet is at 0. A commit may not \
-             reintroduce a family that has been repaired."
-        ));
-    }
-    if destroyed_flags != 0 {
-        failures.push(format!(
-            "{destroyed_flags} real flag(s) destroyed by a collapse; the ratchet is at 0. This \
-             is the count that says how much recall the defect costs — `tools` alone badly \
-             understates it."
-        ));
-    }
-
-    RatchetOutcome {
-        detector: detector.name(),
-        tools,
-        destroyed_flags,
-        self_checks,
-        failures,
-    }
-}
-
-/// `xtask detector self-check`: re-run one detector's own hand-built cases,
-/// or every detector's.
-///
-/// The cheap half of [`ratchet_at_zero`], usable without a `PATH` sweep —
-/// it spawns nothing and reads no fixture, so CI can run it in a second on
-/// every commit while the fleet half only runs where a sweep does.
-pub fn cmd_self_check(detector: Option<&str>) -> anyhow::Result<()> {
-    validate_registry_scopes().map_err(|e| anyhow::anyhow!("{e}"))?;
-    let detectors = match detector {
-        Some(name) => vec![find(name)?],
-        None => registry(),
-    };
-    let mut broken = Vec::new();
-    for d in &detectors {
-        let outcomes = run_self_checks(d.as_ref());
-        println!("{}", "=".repeat(76));
-        println!("detector: {}", d.name());
-        println!("{}", render_self_checks(&outcomes));
-        if outcomes.iter().any(|o| !o.held) {
-            broken.push(d.name());
-        }
-    }
-    if !broken.is_empty() {
-        anyhow::bail!(
-            "self-check failed for: {} — see the FAILED case(s) above",
-            broken.join(", ")
-        );
-    }
-    Ok(())
-}
-
-/// The single-dash long-option split (`crate::single_dash_long`): `-help`
-/// read as `-h` carrying a required value `"elp"`.
-///
-/// The second of the three families sharing `short && !long && value_name`,
-/// and the one spec §13.1's K1 pre-tag is named after. Registered beside
-/// [`BundledShortFlag`] and [`RepeatedCharFlag`] specifically so the
-/// harness's "fired on a tool judged defective of another family" cell can
-/// answer the question the three of them exist to keep honest: whether a
-/// detector for one family is quietly counting another's findings.
-pub(crate) struct SingleDashLong;
-
-/// `single-dash-long`'s declared exclusions. Both are labelled members of the
-/// family that this detector's own conditions refuse, each carrying the
-/// witness token its [`Ground`] recomputes the refusal from.
-const SINGLE_DASH_LONG_EXCLUSIONS: &[Exclusion] = &[
-    Exclusion {
-        tool: "ip",
-        ground: Ground::OptionalBracketedTail {
-            token: crate::single_dash_long::IP_BRACKETED_TOKEN,
-        },
-        note: "ip writes its long options as abbreviation-plus-bracketed-tail \
-               (`-h[uman-readable]`, `-b[atch]`, `-rc[vbuf]`), which the grammar records as an \
-               Optional value — a value spec a human deliberately typed, which this detector's \
-               Required-only fingerprint excludes for the same reason bundling's does",
-    },
-    Exclusion {
-        tool: "sg_emc_trespass",
-        ground: Ground::TailIsNotAnOptionName {
-            token: crate::single_dash_long::SG_EMC_TRESPASS_TOKEN,
-        },
-        note: "its help text glues the layout's own colon onto the flag (`-hr: Set Honor \
-               Reservation bit`), so the tree stores `-h` + \"r:\" and the tail is not a name. \
-               A real miss, and one no tail-shape rule can claim without also admitting every \
-               value spec that leaks punctuation",
-    },
-];
-
-impl Detector for SingleDashLong {
-    fn name(&self) -> &'static str {
-        "single-dash-long"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("single-dash-long")
-    }
-    fn describes(&self) -> &'static str {
-        "an option-table row naming a single-dash long option (`-help`, `-fdump-scos`) parsed as \
-         a one-character short flag carrying the rest of the name as a required value"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::single_dash_long::detect(evidence.raw, evidence.root)
-            .splits
-            .iter()
-            .map(|s| {
-                format!(
-                    "{:?} at {:?} was split into {:?} plus a required value",
-                    s.token, s.path, s.spelling
-                )
-            })
-            .collect()
-    }
-    fn scope(&self) -> Scope {
-        Scope {
-            claim: "option-table-sourced (never synopsis) single-dash tokens whose tail is \
-                    option-name-shaped, at least `single_dash_long::MIN_SWALLOWED_CHARS` long, \
-                    not a repeat of the flag's own letter, and uniformly lowercase. The case \
-                    condition is the load-bearing one and it is not free: it is what keeps the \
-                    entire GCC/Clang glued-value convention out (`cargo -Zscript`, `rpcgen \
-                    -Dname`, `makewhatis -Tutf8`, `perl -Idirectory`, `cc -oOUTFILE`), all of \
-                    them correct parses, and it is equally why an UPPERCASE-led long option is \
-                    knowingly out of reach — no measured signal separates the two",
-            known_exclusions: SINGLE_DASH_LONG_EXCLUSIONS,
-        }
-    }
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        crate::single_dash_long::self_checks()
-    }
-}
-
-pub(crate) struct RepeatedCharFlag;
-
-impl Detector for RepeatedCharFlag {
-    fn name(&self) -> &'static str {
-        "repeated-char-flag"
-    }
-    fn family(&self) -> Option<&'static str> {
-        Some("repeated-char-flag")
-    }
-    fn describes(&self) -> &'static str {
-        "a flag whose required value is its own short character repeated (`-vv` -> `-v` + \"v\"), \
-         in a document that also declares the bare short flag a boolean"
-    }
-    fn hits(&self, evidence: &ToolEvidence<'_>) -> Vec<String> {
-        crate::repeated_char::detect(evidence.raw, evidence.root)
-            .misreads
-            .iter()
-            .map(|m| {
-                format!(
-                    "{:?} at {:?} was read as {:?} carrying its own letter as a value",
-                    m.token, m.path, m.spelling
-                )
-            })
-            .collect()
-    }
-    fn scope(&self) -> Scope {
-        Scope {
-            claim: "repeated-character tokens in a document that ALSO writes the bare short flag \
-                    as a boolean row. That last condition is the whole safety argument and it \
-                    costs real recall: `strace`'s `[-DDD]` and every other synopsis that repeats \
-                    a switch without also writing it alone is out of reach, because the only \
-                    evidence that would admit them is the token's shape — and `lessecho`'s real \
-                    `[-nn]` (a genuine `-n` taking a number) has exactly that shape and is a \
-                    correct parse. No labelled member of this family is excluded",
-            known_exclusions: &[],
-        }
-    }
-    fn self_checks(&self) -> Vec<SelfCheck> {
-        crate::repeated_char::self_checks()
-    }
 }
 
 #[cfg(test)]
